requirements.txt
requests
python-pushover
pyyaml
